refactor:jwt, member, security / add: redis
