+++ study/spring-board/src/main/java/com/kms/springboard/auth/dto/LoginDto.java
@@ -1,7 +1,6 @@
-package com.kms.springboard.member.dto;
+package com.kms.springboard.auth.dto;
 
 import jakarta.validation.constraints.NotBlank;
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
