+++ study/spring-board/src/main/java/com/kms/springboard/auth/jwt/JwtTokenValidator.java
@@ -1,4 +1,4 @@
-package com.kms.springboard.security.jwt;
+package com.kms.springboard.auth.jwt;
 
 
 import lombok.RequiredArgsConstructor;
