+++ study/spring-board/src/main/java/com/kms/springboard/common/Normalizer.java
@@ -1,0 +1,17 @@
+package com.kms.springboard.common;
+
+
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+import org.springframework.stereotype.Component;
+
+import java.util.Locale;
+
+
+@Component
+public class Normalizer {
+
+    public String normalize(String user) {
+        return user==null ? "" : user.trim().toLowerCase(Locale.ROOT);
+    }
+}
