+++ study/spring-board/src/main/java/com/kms/springboard/member/repository/MemberRepository.java
@@ -3,7 +3,7 @@ package com.kms.springboard.member.repository;
 import com.kms.springboard.member.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
+
 import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
