+++ study/spring-board/src/main/java/com/kms/springboard/member/service/MemberService.java
@@ -1,20 +1,14 @@
 package com.kms.springboard.member.service;
 
 
-import com.kms.springboard.member.dto.LoginDto;
 import com.kms.springboard.member.dto.MemberDto;
 import com.kms.springboard.member.entity.MemberEntity;
-import com.kms.springboard.security.jwt.dto.JwtTokenResponse;
-import lombok.RequiredArgsConstructor;
 
 import java.util.Optional;
 
 
 public interface MemberService {
 
-    MemberEntity save(MemberEntity member);
     MemberEntity saveDto(MemberDto memberDto);
-    boolean isLogin(LoginDto loginDto);
     Optional <MemberEntity> findByUserId(String userId);
-    JwtTokenResponse login(LoginDto loginDto);
 }
