+++ study/spring-board/src/main/java/com/kms/springboard/member/service/MemberServiceImpl.java
@@ -1,19 +1,15 @@
 package com.kms.springboard.member.service;
 
-import com.kms.springboard.security.jwt.JwtTokenGenerator;
-import com.kms.springboard.member.dto.LoginDto;
+import com.kms.springboard.common.Normalizer;
 import com.kms.springboard.member.dto.MemberDto;
 import com.kms.springboard.member.entity.MemberEntity;
 import com.kms.springboard.member.repository.MemberRepository;
-import com.kms.springboard.security.jwt.dto.JwtTokenResponse;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Locale;
 import java.util.Optional;
 
 @RequiredArgsConstructor
@@ -23,22 +19,12 @@ public class MemberServiceImpl implements MemberService {
 
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
-    private final JwtTokenGenerator jwtTokenGenerator;
-
-    @Override
-    public MemberEntity save(MemberEntity member) {
-        String pw = member.getPassword();
-        if(pw == null || !pw.startsWith("$2")){
-            throw new IllegalArgumentException("Password must be encoded before save()");
-        }
-        return memberRepository.save(member);
-
-    }
+    private final Normalizer normalizer;
 
     @Override
     public MemberEntity saveDto(MemberDto memberDto) {
-        final var normalizedUserId = memberDto.getUserId().trim().toLowerCase(Locale.ROOT);
-        final var normalizedEmail = memberDto.getEmail().trim().toLowerCase(Locale.ROOT);
+        final var normalizedUserId = normalizer.normalize(memberDto.getUserId());
+        final var normalizedEmail =  normalizer.normalize(memberDto.getEmail());
         if(memberRepository.existsByUserId(normalizedUserId)) {
             throw new IllegalStateException("이미 존재하는 아이디입니다");
         }
@@ -56,23 +42,7 @@ public class MemberServiceImpl implements MemberService {
         return memberRepository.save(member);
     }
 
-    @Override
-    @Transactional(readOnly = true)
-    public boolean isLogin(LoginDto loginDto) {
-        String userId = loginDto.getUserId();
-        String password = loginDto.getPassword();
 
-        if(userId == null || userId.isBlank() || password == null || password.isBlank()) {
-            return false;
-        }
-        final String normalized = userId == null ? "" : userId.trim().toLowerCase(Locale.ROOT);
-        return memberRepository.findByUserId(normalized)
-                .map(member-> passwordEncoder.matches(password,member.getPassword()))
-                        .orElse(false);
-
-
-
-    }
 
     @Override
     @Transactional(readOnly = true)
@@ -80,16 +50,6 @@ public class MemberServiceImpl implements MemberService {
         return memberRepository.findByUserId(userId);
     }
 
-
-    @Override
-    @Transactional(readOnly = true)
-    public JwtTokenResponse login(LoginDto loginDto) {
-        if(!isLogin(loginDto)){
-            throw new EntityNotFoundException("아이디 또는 비밀번호가 잘못되었습니다");
-        }
-        final String normalizedUserId = loginDto.getUserId().trim().toLowerCase(Locale.ROOT);
-        return jwtTokenGenerator.generateToken(normalizedUserId);
-    }
 
 
 }
